+++ blueprint/implementations/billing/base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/implementation-billing-base
 
+## 0.8.16
+
+### Patch Changes
+
+- small nits
+- Updated dependencies
+  - @forklaunch/interfaces-billing@0.8.16
+
 ## 0.8.15
 
 ### Patch Changes